+++ src/main/java/by/golik/dealerstat/entity/AbstractEntity.java
@@ -21,7 +21,7 @@ public abstract class AbstractEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected Long id;
+    protected int id;
 
     @CreatedDate
     @Temporal(TemporalType.TIMESTAMP)
